+++ js/constants.js
@@ -1,0 +1,22 @@
+const COUNT_OF_AD = 10;
+const TYPE_OF_OFFER = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel'
+];
+const FEATURES_OF_OFFER = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner'
+];
+
+export {
+  COUNT_OF_AD,
+  TYPE_OF_OFFER,
+  FEATURES_OF_OFFER
+};
