--- js/create-objects.js
@@ -1,6 +1,0 @@
-import {
-  CreateArraySimilarAdsNearby,
-  COUNT_OF_AD
-} from './data.js';
-
-CreateArraySimilarAdsNearby(COUNT_OF_AD);
