+++ js/data.js
@@ -1,10 +1,15 @@
 import {
   getRandomIntegerNumberFromRange,
   getRandomNotIntegerNumberFromRange,
-  getRandomArrayElement
+  getRandomArrayElement,
+  getNewSetOfValues
 } from './util.js';
+import {
+  COUNT_OF_AD,
+  TYPE_OF_OFFER,
+  FEATURES_OF_OFFER
+} from './constants.js';
 
-const COUNT_OF_AD = 10;
 const CoordinatesOfOffer = {
   MAXLAT : 35.70000,
   MINLAT : 35.65000,
@@ -12,28 +17,21 @@ const CoordinatesOfOffer = {
   MINLNG : 139.70000,
   NUMBER_OF_DIGITS_AFTER_COMMA : 5
 };
-const priceOfOffer = {
+const PriceOfOffer = {
   MIN_PRICE : 1000,
   MAX_PRICE : 10000
 };
-const roomsOfOffer = {
+const RoomsOfOffer = {
   MIN_COUNT_OF_ROOMS : 1,
   MAX_COUNT_OF_ROOMS : 3
 };
-const guestsOfOffer =  {
+const GuestsOfOffer =  {
   MIN_COUNT_OF_GUESTS : 1,
   MAX_COUNT_OF_GUESTS: 6
 };
 const TITLE_OF_OFFER = [
   'My flat for you',
   'Be happy here',
-];
-const TYPE_OF_OFFER = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-  'hotel'
 ];
 const CHECIN_OF_OFFER = [
   '12:00',
@@ -45,14 +43,6 @@ const CHECKOUT_OF_OFFER = [
   '13:00',
   '14:00'
 ];
-const FEATURES_OF_OFFER = [
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner'
-];
 const DESCRIPTION_OF_OFFER = [
   'Very good',
   'Good'
@@ -62,18 +52,11 @@ const PHOTOS_OF_OFFER = [
   'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
   'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
 ];
-
 const getAvatars = (count) => {
   const avatars = new Array(count).fill('').map((_, i) => `img/avatars/user${String(i+1).padStart(2, '0')}.png`);
   return avatars;
 };
 const AVATARS_OF_AUTHOR = getAvatars(COUNT_OF_AD);
-
-const getNewSetOfValues = (values) => {
-  const setOfValues = [...values];
-  setOfValues.sort(() => Math.random() - 0.5);
-  return setOfValues.slice(0, getRandomIntegerNumberFromRange(1, values.length));
-};
 
 const createAds = (number) => {
   const latOfOffer = getRandomNotIntegerNumberFromRange(CoordinatesOfOffer.MINLAT, CoordinatesOfOffer.MAXLAT, CoordinatesOfOffer.NUMBER_OF_DIGITS_AFTER_COMMA);
@@ -85,10 +68,10 @@ const createAds = (number) => {
     offer : {
       title : getRandomArrayElement(TITLE_OF_OFFER),
       address : `${latOfOffer}, ${lngOfOffer}`,
-      price : getRandomIntegerNumberFromRange(priceOfOffer.MIN_PRICE, priceOfOffer.MAX_PRICE),
+      price : getRandomIntegerNumberFromRange(PriceOfOffer.MIN_PRICE, PriceOfOffer.MAX_PRICE),
       type : getRandomArrayElement(TYPE_OF_OFFER),
-      rooms : getRandomIntegerNumberFromRange(roomsOfOffer.MIN_COUNT_OF_ROOMS, roomsOfOffer.MAX_COUNT_OF_ROOMS),
-      guests : getRandomIntegerNumberFromRange(guestsOfOffer.MIN_COUNT_OF_GUESTS, guestsOfOffer.MAX_COUNT_OF_GUESTS),
+      rooms : getRandomIntegerNumberFromRange(RoomsOfOffer.MIN_COUNT_OF_ROOMS, RoomsOfOffer.MAX_COUNT_OF_ROOMS),
+      guests : getRandomIntegerNumberFromRange(GuestsOfOffer.MIN_COUNT_OF_GUESTS, GuestsOfOffer.MAX_COUNT_OF_GUESTS),
       checkin : getRandomArrayElement(CHECIN_OF_OFFER),
       checkout : getRandomArrayElement(CHECKOUT_OF_OFFER),
       features : getNewSetOfValues(FEATURES_OF_OFFER),
@@ -102,9 +85,6 @@ const createAds = (number) => {
   };
 };
 
-const CreateArraySimilarAdsNearby = (count) => new Array(count).fill('').map((_, i) => createAds(i));
+const createOffers = (count) => new Array(count).fill('').map((_, i) => createAds(i));
 
-export {
-  CreateArraySimilarAdsNearby,
-  COUNT_OF_AD
-};
+export {createOffers};
