+++ js/main.js
@@ -1,1 +1,4 @@
-import './create-objects.js';
+import {createOffers} from './data.js';
+import {COUNT_OF_AD} from './constants.js';
+
+createOffers(COUNT_OF_AD);
